avoid cycles in imports
